Refactored thoughts_callback to autogen_utils.py
