+++ experiments/Retrosynthesis/main.py
@@ -68,38 +68,11 @@ if __name__ == "__main__":
         runner = GeminiClient(experiment_type=myexperiment, api_key=client_key)
     elif args.client == "autogen":
         from charge.clients.autogen import AutoGenClient
+        from charge.clients.autogen_utils import thoughts_callback
 
         (model, backend, API_KEY, kwargs) = AutoGenClient.configure(
             args.model, args.backend
         )
-
-        def thoughts_callback(assistant_message):
-            # print("In callback:", assistant_message)
-            if assistant_message.type == "UserMessage":
-                print(f"User: {assistant_message.content}")
-            elif assistant_message.type == "AssistantMessage":
-
-                if assistant_message.thought is not None:
-                    print(f"Model thought: {assistant_message.thought}")
-                if isinstance(assistant_message.content, list):
-                    for item in assistant_message.content:
-                        if hasattr(item, "name") and hasattr(item, "arguments"):
-                            print(
-                                f"Function call: {item.name} with args {item.arguments}"
-                            )
-                        else:
-                            print(f"Model: {item}")
-            elif assistant_message.type == "FunctionExecutionResultMessage":
-
-                for result in assistant_message.content:
-                    if result.is_error:
-                        print(
-                            f"Function {result.name} errored with output: {result.content}"
-                        )
-                    else:
-                        print(f"Function {result.name} returned: {result.content}")
-            else:
-                print("Model: ", assistant_message.message.content)
 
         runner = AutoGenClient(
             experiment_type=myexperiment,
